requirements.txt
Pillow~=9.1.0 # Image handling
requests~=2.28.1 # HTTP Requests
python-dotenv~=0.20.0 # Load .env files
